+++ R/Function_SinkFill.r
@@ -6,7 +6,7 @@ SinkFill <-
         if ( missing( DEM ) || !is( DEM, "RasterLayer" ) )
         { 
             stop("Invalid type of DEM") 
-        }else if ( xres( DEM ) == yres( DEM ) )
+        }else if ( all.equal( xres( DEM ), yres( DEM ) ) )
         {
             # Extract basic infomation of a DEM
             res     <- xres( DEM )